fix window DST range and route helper
